Harden adb-remount-test.sh
I'm getting this error:

  [  FAILED  ] /cache/overlay wipe

Check file path existence before removing.
If /cache is not mounted, in which case it would be RO, then don't try
to do `rm -rf /cache/overlay` as it would fail, albeit the "-f" flag.

Bug: 178683776
Test: adb-remount-test.sh on GSI + Pixel

+++ fs_mgr/tests/adb-remount-test.sh
@@ -1656,8 +1656,10 @@ echo "${GREEN}[       OK ]${NORMAL} remount command works from setup" >&2
 # This also saves a lot of 'noise' from the command doing a mkfs on backing
 # storage and all the related tuning and adjustment.
 for d in ${OVERLAYFS_BACKING}; do
-  adb_su rm -rf /${d}/overlay </dev/null ||
-    die "/${d}/overlay wipe"
+  if adb_su ls -d /${d}/overlay </dev/null >/dev/null 2>/dev/null; then
+    adb_su rm -rf /${d}/overlay </dev/null ||
+      die "/${d}/overlay wipe"
+  fi
 done
 adb_reboot &&
   adb_wait ${ADB_WAIT} ||
